move GeneratedData struct and PlaceholderMsg to same package under common
