+++ builder/docker/step_set_generated_data_test.go
@@ -4,14 +4,14 @@ import (
 	"context"
 	"testing"
 
-	"github.com/hashicorp/packer/builder"
+	"github.com/hashicorp/packer/common/packerbuilderdata"
 	"github.com/hashicorp/packer/helper/multistep"
 )
 
 func TestStepSetGeneratedData_Run(t *testing.T) {
 	state := testState(t)
 	step := new(StepSetGeneratedData)
-	step.GeneratedData = &builder.GeneratedData{State: state}
+	step.GeneratedData = &packerbuilderdata.GeneratedData{State: state}
 	driver := state.Get("driver").(*MockDriver)
 	driver.Sha256Result = "80B3BB1B1696E73A9B19DEEF92F664F8979F948DF348088B61F9A3477655AF64"
 	state.Put("image_id", "12345")
@@ -33,7 +33,7 @@ func TestStepSetGeneratedData_Run(t *testing.T) {
 
 	// Image ID not implement
 	state = testState(t)
-	step.GeneratedData = &builder.GeneratedData{State: state}
+	step.GeneratedData = &packerbuilderdata.GeneratedData{State: state}
 	driver = state.Get("driver").(*MockDriver)
 	notImplementedMsg := "ERR_IMAGE_SHA256_NOT_FOUND"
