+++ common/packerbuilderdata/generated_data.go
@@ -1,6 +1,11 @@
-package builder
+package packerbuilderdata
 
 import "github.com/hashicorp/packer/helper/multistep"
+
+// This is used in the BasicPlaceholderData() func in the packer/provisioner.go
+// To force users to access generated data via the "generated" func.
+const PlaceholderMsg = "To set this dynamically in the Packer template, " +
+	"you must use the `build` function"
 
 // GeneratedData manages variables exported by a builder after
 // it started. It uses the builder's multistep.StateBag internally, make sure it
