+++ common/packerbuilderdata/generated_data_test.go
@@ -1,4 +1,4 @@
-package builder
+package packerbuilderdata
 
 import (
 	"testing"
